backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
Flask-Limiter==3.8.0
opencv-python==4.10.0.84
face-recognition==1.3.0
numpy==1.26.4
pymongo==4.8.0
mongomock==4.1.2
python-dotenv==1.0.1
PyJWT==2.9.0
gunicorn==22.0.0
redis==5.0.8
sentry-sdk[flask]==2.13.0
